re-organize HMC's different elements
