+++ blackjax/inference/proposals.py
@@ -29,11 +29,14 @@ References
 """
 from typing import Callable, Dict, List, NamedTuple, Tuple, Union
 
+import blackjax.inference.termination as termination
+import blackjax.inference.trajectory as trajectory
+import blackjax.inference.mh_step as mh_step
 import jax
 import jax.numpy as jnp
 from blackjax.inference.integrators import IntegratorState
-import blackjax.inference.trajectory as trajectory
-import blackjax.inference.termination as termination
+from blackjax.inference.mh_step import Proposal
+from blackjax.inference.trajectory import Trajectory
 
 __all__ = ["hmc"]
 
@@ -161,8 +164,7 @@ def hmc(
     def generate(
         rng_key, initial_state: IntegratorState
     ) -> Tuple[IntegratorState, ProposalInfo]:
-        """Generate a new chain state.
-        """
+        """Generate a new chain state."""
         end_state = integrate_trajectory(initial_state)
         proposal_state = flip_momentum(end_state)
         new_state, info = maybe_update_proposal(rng_key, initial_state, proposal_state)
@@ -181,19 +183,18 @@ def iterative_nuts(
 ):
     """Iterative NUTS proposal."""
 
-    proposal_fn = proposal_generator(kinetic_energy, divergence_threshold)
-
+    proposal_fn = mh_step.proposal_generator(kinetic_energy, divergence_threshold)
     (
         new_criterion_state,
         update_criterion_state,
         is_criterion_met,
     ) = termination.iterative_uturn_numpyro(uturn_check_fn)
 
-    trajectory_integrator = dynamic_integration(
+    trajectory_integrator = trajectory.dynamic_progressive_integration(
         integrator, proposal_fn, update_criterion_state, is_criterion_met
     )
 
-    expand, do_keep_expanding = dynamic_multiplicative_expansion(
+    expand, do_keep_expanding = trajectory.dynamic_multiplicative_expansion(
         trajectory_integrator,
         uturn_check_fn,
         step_size,
@@ -218,399 +219,3 @@ def iterative_nuts(
         return proposal.state, None
 
     return propose
-
-
-# -------------------------------------------------------------------
-#                            TRAJECTORY
-# -------------------------------------------------------------------
-
-# Q: Can I just create a class with `append` and `__add__` methods?
-class Trajectory(NamedTuple):
-    leftmost_state: IntegratorState
-    rightmost_state: IntegratorState
-    momentum_sum: PyTree
-
-
-def append_to_trajectory(direction, trajectory, state):
-    """Append a state to the trajectory to form a new trajectory."""
-    leftmost_state, rightmost_state = jax.lax.cond(
-        direction > 0,
-        trajectory,
-        lambda trajectory: (trajectory.leftmost_state, state),
-        trajectory,
-        lambda trajectory: (state, trajectory.rightmost_state),
-    )
-    momentum_sum = jax.tree_util.tree_multimap(
-        jnp.add, trajectory.momentum_sum, state.momentum
-    )
-    return Trajectory(leftmost_state, rightmost_state, momentum_sum)
-
-
-def merge_trajectories(
-    direction: int, trajectory: Trajectory, new_trajectory: Trajectory
-) -> Trajectory:
-    """Merge two trajectories to form a new trajectory."""
-    leftmost_state, rightmost_state = jax.lax.cond(
-        direction > 0,
-        trajectory,
-        lambda trajectory: (
-            trajectory.leftmost_state,
-            new_trajectory.rightmost_state,
-        ),
-        trajectory,
-        lambda trajectory: (
-            new_trajectory.leftmost_state,
-            trajectory.rightmost_state,
-        ),
-    )
-    momentum_sum = jax.tree_util.tree_multimap(
-        jnp.add, trajectory.momentum_sum, new_trajectory.momentum_sum
-    )
-    return Trajectory(leftmost_state, rightmost_state, momentum_sum)
-
-
-# -------------------------------------------------------------------
-#                   TRAJECTORY EXPANSION
-# -------------------------------------------------------------------
-
-
-def dynamic_multiplicative_expansion(
-    integrator: Callable,
-    uturn_check_fn: Callable,
-    step_size: float,
-    max_tree_depth: int = 10,
-    rate: int = 2,
-) -> Tuple[Callable, Callable]:
-    """Sample a trajectory and update the proposal sequentially
-    until the termination criterion is met.
-
-    The trajectory is sampled with the following procedure:
-    1. Pick a direction at random
-    2. Integrate `num_step` steps in this direction
-    3. If integration stopped prematurely, do not update proposal
-    4. Else if trajectory is performing a U-turn, return proposal
-    5. Else update proposal, `num_steps = num_steps ** rate` and repeat from (1).
-
-    Parameters
-    ----------
-    integrator
-    uturn_check_fn
-        Function used to check the U-Turn criterion.
-    step_size
-        The step size used by the trajectory integrator.
-    max_tree_depth
-        The maximum number of trajectory expansions until the proposal is
-        returned.
-    rate
-        The rate of the geometrical expansion. Typically 2 in NUTS, hence
-        the mentions to binary trees.
-
-    """
-
-    def do_keep_expanding(expansion_state) -> bool:
-        """Determine whether we need to keep expanding the trajectory."""
-        _, trajectory, _, _, terminated_early, depth = expansion_state
-        is_turning = uturn_check_fn(
-            trajectory.leftmost_state.momentum,
-            trajectory.rightmost_state.momentum,
-            trajectory.momentum_sum,
-        )
-        return (depth < max_tree_depth) & ~terminated_early & ~is_turning
-
-    def expand(expansion_state):
-        """Expand the current trajectory.
-
-        At each step we draw a direction at random, build a subtrajectory starting
-        from the leftmost or rightmost point of the current trajectory that is
-        twice as long as the current trajectory.
-
-        Once that is done, possibly update the current proposal with that of
-        the subtrajectory.
-
-        """
-        # Q: Should this function be aware of all the elements that need to
-        # be passed downstream?
-        rng_key, trajectory, proposal, termination_state, _, depth = expansion_state
-        rng_key, direction_key = jax.random.split(rng_key, 2)
-
-        # create new subtrajectory that is twice as long as the current
-        # trajectory.
-        direction = jnp.where(jax.random.bernoulli(direction_key), 1, -1)
-        start_state = jax.lax.cond(
-            direction > 0,
-            trajectory,
-            lambda trajectory: trajectory.rightmost_state,
-            trajectory,
-            lambda trajectory: trajectory.leftmost_state,
-        )
-        (
-            new_proposal,
-            new_trajectory,
-            termination_state,
-            terminated_early,
-        ) = integrator(
-            rng_key, start_state, direction, termination_state, rate ** depth, step_size
-        )
-
-        # merge the freshly integrated trajectory to the current trajectory
-        new_trajectory = merge_trajectories(direction, trajectory, new_trajectory)
-
-        # update the proposal
-        # we reject proposals coming from diverging or turning subtrajectories
-        maybe_updated_proposal = jax.lax.cond(
-            terminated_early,
-            proposal,
-            lambda x: x,
-            (rng_key, proposal, new_proposal),
-            lambda x: progressive_biased_sampling(*x),
-        )
-
-        return (
-            rng_key,
-            new_trajectory,
-            maybe_updated_proposal,
-            termination_state,
-            terminated_early,
-            depth + 1,
-        )
-
-    return expand, do_keep_expanding
-
-
-def dynamic_integration(
-    integrator: Callable,
-    proposal_generator: Callable,
-    update_termination: Callable,
-    is_criterion_met: Callable,
-):
-    """Integrate a trajectory and update the proposal sequentially
-    until the termination criterion is met.
-
-    Parameters
-    ----------
-    integrator
-        The symplectic integrator used to integrate the hamiltonian trajectory.
-    termination_criterion
-        The criterion used to stop the sampling. This function generates a function that
-        initializes the state, a function that updates it and a function that determines
-        whether the criterion is met.
-    progressive_sample
-        The sampler used to update the proposal as the trajectory is being sampled. Generates
-        a function that initializes the proposal state and one that updates it.
-
-    """
-
-    def integrate(
-        rng_key: jax.numpy.DeviceArray,
-        initial_state: IntegratorState,
-        direction: int,
-        termination_state,
-        max_num_steps: int,
-        step_size,
-    ):
-        """Integrate the trajectory starting from `initial_state` and update
-        the proposal sequentially until the termination criterion is met.
-
-        Parameters
-        ----------
-        rng_key
-            Key used by JAX's random number generator.
-        initial_state
-            The initial state from which we start expanding the trajectory.
-        direction int in {-1, 1}
-            The direction in which to expand the trajectory.
-        max_num_steps
-            The maximum number of integration steps. The expansion will stop
-            when this number is reached if the termination criterion has not
-            been met.
-        step_size
-            The step size of the symplectic integrator.
-
-        """
-
-        # QUESTION: Do we only signal a divergence if the proposal is issued from
-        # a divergent transition (current) or anytime a divergence is detected?
-        def do_keep_integrating(expansion_state):
-            """Decide whether we should continue integrating the trajectory"""
-            _, state, trajectory, proposal, termination_state, step = expansion_state
-            has_terminated = is_criterion_met(termination_state, trajectory, state)
-            is_diverging = proposal.is_diverging
-            return (step < max_num_steps) & ~has_terminated & ~is_diverging
-
-        def add_one_state(expansion_state):
-            """We need to compute the delta energy between the last and the new state
-            proposed. Tree weight is taken to be energy_current - energy_new
-            """
-            (
-                rng_key,
-                state,
-                trajectory,
-                proposal,
-                termination_state,
-                step,
-            ) = expansion_state
-            _, rng_key = jax.random.split(rng_key)
-
-            new_state = integrator(state, direction * step_size)
-            new_trajectory = append_to_trajectory(direction, trajectory, new_state)
-
-            # Q: couldn't we pass `state` instead of `(trajectory, direction)`?
-            new_proposal = proposal_generator(direction, new_state, trajectory)
-            maybe_updated_proposal = progressive_uniform_sampling(
-                rng_key, proposal, new_proposal
-            )
-
-            new_termination_state = update_termination(
-                termination_state, new_trajectory, new_state, step
-            )
-
-            return (
-                rng_key,
-                new_state,
-                new_trajectory,
-                maybe_updated_proposal,
-                new_termination_state,
-                step + 1,
-            )
-
-        initial_integration_state = (
-            rng_key,
-            initial_state,
-            Trajectory(initial_state, initial_state, initial_state.momentum),
-            Proposal(initial_state, 0.0, False),
-            termination_state,
-            0,
-        )
-
-        _, _, trajectory, proposal, termination_state, step = jax.lax.while_loop(
-            do_keep_integrating, add_one_state, initial_integration_state
-        )
-
-        return proposal, trajectory, termination_state, (step < max_num_steps)
-
-    return integrate
-
-
-# -------------------
-#     PROPOSALS
-# -------------------
-
-
-class Proposal(NamedTuple):
-    """Proposal for the next chain step.
-
-    state:
-        The trajectory state corresponding to this proposal.
-    weight:
-        The logarithm of the sum of the canonical densities of each state
-        :math:`e^{-H(z)}` along the trajectory.
-    is_diverging
-        Whether a divergence was observed when making one step.
-    """
-
-    state: IntegratorState
-    log_weight: float
-    is_diverging: bool
-
-
-def proposal_generator(kinetic_energy: Callable, divergence_threshold: float):
-    def generate(direction, state, trajectory):
-        """Generate a new proposal from a trajectory state.
-
-        Note
-        ----
-        We can optimize things a bit by sacrificing in memory and code
-        complexity: the kinetic energy and the previous states can be carried
-        around instead of being recomputed each time.
-
-        Parameters
-        ----------
-        direction:
-            The direction in which the trajectory was integrated to obtain the
-            new state.
-        state:
-            The new state.
-        trajectory:
-            The trajectory before the state was added.
-
-        """
-        previous_state = jax.lax.cond(
-            direction > 0,
-            trajectory,
-            lambda x: x.rightmost_state,
-            trajectory,
-            lambda x: x.leftmost_state,
-        )
-
-        energy = previous_state.potential_energy + kinetic_energy(
-            previous_state.position, previous_state.momentum
-        )
-        new_energy = state.potential_energy + kinetic_energy(
-            state.position, state.momentum
-        )
-
-        delta_energy = energy - new_energy
-        delta_energy = jnp.where(jnp.isnan(delta_energy), -jnp.inf, delta_energy)
-        is_diverging = jnp.abs(delta_energy) > divergence_threshold
-
-        # The log-weight of the new proposal is equal to H(z) - H(z_new)?
-        log_weight = delta_energy
-
-        return Proposal(
-            state,
-            log_weight,
-            is_diverging,
-        )
-
-    return generate
-
-
-def progressive_uniform_sampling(rng_key, proposal, new_proposal):
-    """Generate a new proposal.
-
-    To avoid keeping the entire trajectory in memory, we only memorize the
-    extreme points and the point that will currently be proposed as a sample.
-    Progressive sampling updates this proposal as the trajectory is being
-    built. This is scheme is equivalent to drawing a sample uniformly at random
-    from the final trajectory.
-
-    """
-    p_accept = jax.scipy.special.expit(new_proposal.log_weight - proposal.log_weight)
-    do_accept = jax.random.bernoulli(rng_key, p_accept)
-
-    updated_proposal = Proposal(
-        new_proposal.state,
-        jnp.logaddexp(proposal.log_weight, new_proposal.log_weight),
-        new_proposal.is_diverging,
-    )
-
-    return jax.lax.cond(
-        do_accept, lambda _: updated_proposal, lambda _: proposal, operand=None
-    )
-
-
-def progressive_biased_sampling(rng_key, proposal, new_proposal):
-    """Generate a new proposal.
-
-    To avoid keeping the entire trajectory in memory, we only memorize the
-    extreme points and the point that will currently be proposed as a sample.
-    Progressive sampling updates this proposal as the trajectory is being
-    built.
-
-    Unlike uniform sampling, biased sampling favors new proposals. It thus
-    biases the transition away from the trajectory's initial state.
-
-    """
-    p_accept = jnp.exp(new_proposal.log_weight - proposal.log_weight)
-    p_accept = jnp.clip(p_accept, a_max=1.0)
-    do_accept = jax.random.bernoulli(rng_key, p_accept)
-
-    updated_proposal = Proposal(
-        new_proposal.state,
-        jnp.logaddexp(proposal.log_weight, new_proposal.log_weight),
-        new_proposal.is_diverging,
-    )
-
-    return jax.lax.cond(
-        do_accept, lambda _: updated_proposal, lambda _: proposal, operand=None
-    )
